% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtcars_missing}
\alias{mtcars_missing}
\title{Missing Airline Delays (US)}
\format{
A data frame with columns:
\describe{
\item{car}{The variable is character. }
\item{mpg}{The variable is numeric. Miles/(US) gallon}
\item{cyl}{The variable is integer. Number of cylinders}
\item{disp}{The variable is numeric. Displacement (cu.in.)}
\item{hp}{The variable is integer. Gross horsepower}
\item{drat}{The variable is numeric. Rear axle ratio}
\item{wt}{The variable is numeric. Weight (1000 lbs)}
\item{qsec}{The variable is numeric. 1/4 mile time}
\item{vs}{The variable is integer. Engine (0 = V-shaped, 1 = straight)}
\item{am}{The variable is integer. Transmission (0 = automatic, 1 = manual)}
\item{gear}{The variable is integer. Number of forward gears}
\item{carb}{The variable is integer. Number of carburetors}
}
}
\source{
\url{datasets::mtcars, Henderson and Velleman (1981), Building multiple regression models interactively. Biometrics, 37, 391–411.}
}
\usage{
mtcars_missing
}
\description{
Monthly Airline Delays by Airport for US Flights, 2003-2016
}
\examples{
\dontrun{
mtcars_missing
}
}
\keyword{datasets}
